resoved issues

+++ resources/js/gallery_filter_search.js
@@ -7,51 +7,52 @@ $(function() {
         }
     });
 
-    $(document).ready(function() {
-        fetch_data();
-        $('#gallerySearch').keyup(function () {
+    if($('#tableGallery').length) {
+        $(document).ready(function() {
             fetch_data();
-        });
-        $("#filterGallery").change(function () {
-            fetch_data();
-        });
-        $("#page").change(function () {
-            fetch_data();
-        });
+            $('#gallerySearch').keyup(function () {
+                fetch_data();
+            });
+            $("#filterGallery").change(function () {
+                fetch_data();
+            });
+            $("#page").change(function () {
+                fetch_data();
+            });
 
-        $(document).on('click', '.delete-action-gallery', function() {
-            const galleryId = $(this).data('id');
-            Swal.fire({
-                title: 'Are you sure?',
-                text: "You won't be able to revert this!",
-                icon: 'warning',
-                showCancelButton: true,
-                confirmButtonColor: '#3085D6',
-                cancelButtonColor: '#d33',
-                confirmButtonText: 'Yes, delete it!'
-            }).then((result) => {
-                if (result.isConfirmed) {
-                    deleteGallery(galleryId);
-                    Swal.fire(
-                        'Deleted!',
-                        'Profession has been deleted.',
-                        'success'
-                    )
-                }
-            })
+            $(document).on('click', '.delete-action-gallery', function() {
+                const galleryId = $(this).data('id');
+                Swal.fire({
+                    title: 'Are you sure?',
+                    text: "You won't be able to revert this!",
+                    icon: 'warning',
+                    showCancelButton: true,
+                    confirmButtonColor: '#3085D6',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: 'Yes, delete it!'
+                }).then((result) => {
+                    if (result.isConfirmed) {
+                        deleteGallery(galleryId);
+                        Swal.fire(
+                            'Deleted!',
+                            'Profession has been deleted.',
+                            'success'
+                        )
+                    }
+                })
 
+            });
         });
-    });
+    }
 
     let xhr = null;
-    function fetch_data(page, id) {
+    function fetch_data(page) {
         xhr = $.ajax({
             data: {
                 search: $("#gallerySearch").val(),
                 filter: $("#filterGallery").val(),
                 perPage: $("#page").val(),
                 page: page,
-                deleteId: id
             },
             url: "/admin/galleries",
             method: "GET",
@@ -85,11 +86,11 @@ $(function() {
 
     function buildTable(response) {
         if((response.data).length === 0) {
-            $('#tableGallery').hide();
+            $('#gallery_block').hide();
             $('#noGallery').html("No gallery found");
         }
         else {
-            $('#tableGallery').show();
+            $('#gallery_block').show();
             $('#noGallery').empty();
             $('#galleryBody').empty();
             $.each(response.data, function (key, value) {
